CI: audit fixtures against the ABI 0.2 capability set; mode_t is u16 on macOS
The fixture audit still enforced the pre-0.2 rule that a plugin imports no
WASI. Since the plugin platform, WASI p2 is the model, so it now allows the
sicompass interfaces and the WASI interfaces the host links (plugin_abi's
baseline and TCP), audits the released plugin fixtures too, and still refuses
ip-name-lookup, which the host never links.

The desktop stat test compared st_mode with libc::S_IFMT, a u16 on macOS.

+++ src/sicompass/src/wasm_host/desktop.rs
@@ -622,7 +622,12 @@ mod tests {
         assert!(f.modified > 0);
 
         let l = s.stat(root.join("link").to_string_lossy().into_owned()).unwrap();
-        assert_eq!(l.mode.unwrap() & libc::S_IFMT, libc::S_IFLNK, "the link, not its target");
+        // `mode_t` is a u16 on macOS.
+        assert_eq!(
+            l.mode.unwrap() & u32::from(libc::S_IFMT),
+            u32::from(libc::S_IFLNK),
+            "the link, not its target"
+        );
         assert!(s.stat("/etc/passwd".to_owned()).is_err(), "outside the grants");
     }
